Create LICENSE
Disclaimer : This Article/Repository does not mean to violate copyrights or grabbing contents of someone else work or make offense to anyone else. This article will help victims of WP-VCD malware attack. This article is not made to gain merits on behalf of someone. Credits already mentioned above for their great work. This article is here to help people who were/are infected for Wp-vcd malware.